standardize language swicher implementation
