+++ packages/frontend/editor-ui/src/plugins/i18n/locales/zh.json
@@ -2837,6 +2837,8 @@
 	"settings.personal.theme.systemDefault": "System default",
 	"settings.personal.theme.light": "Light theme",
 	"settings.personal.theme.dark": "Dark theme",
+	"settings.language.english": "English",
+	"settings.language.chinese": "中文",
 	"settings.personal.mfa.button.disabled.infobox": "Two-factor authentication is currently disabled.",
 	"settings.personal.mfa.button.enabled.infobox": "Two-factor authentication is currently enabled.",
 	"settings.personal.mfa.button.enabled": "Enable 2FA",
